+++ plugin/src/main/scala/com/stratio/cassandra/lucene/index/FSIndex.scala
@@ -17,7 +17,6 @@ package com.stratio.cassandra.lucene.index
 
 import java.nio.file.Path
 
-import com.stratio.cassandra.lucene.util.Logging
 import org.apache.cassandra.io.util.FileUtils
 import org.apache.lucene.analysis.Analyzer
 import org.apache.lucene.document.Document
@@ -43,8 +42,7 @@ class FSIndex(
     refreshSeconds: Double,
     ramBufferMB: Int,
     maxMergeMB: Int,
-    maxCachedMB: Int)
-  extends Logging {
+    maxCachedMB: Int) {
 
   private[this] var mergeSort: Sort = _
   private[this] var fields: java.util.Set[String] = _
@@ -100,7 +98,6 @@ class FSIndex(
     * @param document the document to be added
     */
   def upsert(term: Term, document: Document) {
-    logger.debug(s"Indexing $document with term $term in $name")
     writer.updateDocument(term, document)
   }
 
@@ -109,7 +106,6 @@ class FSIndex(
     * @param term the term identifying the documents to be deleted
     */
   def delete(term: Term) {
-    logger.debug(s"Deleting $term from $name")
     writer.deleteDocuments(term)
   }
 
@@ -118,7 +114,6 @@ class FSIndex(
     * @param query the query identifying the documents to be deleted
     */
   def delete(query: Query) {
-    logger.debug(s"Deleting $query from $name")
     writer.deleteDocuments(query)
   }
 
@@ -126,13 +121,11 @@ class FSIndex(
   def truncate() {
     writer.deleteAll()
     writer.commit()
-    logger.info(s"Truncated $name")
   }
 
   /** Commits the pending changes. */
   def commit() {
     writer.commit()
-    logger.debug(s"Committed $name")
   }
 
   /** Commits all changes to the index, waits for pending merges to complete, and closes all
@@ -143,13 +136,11 @@ class FSIndex(
     manager.close()
     writer.close()
     directory.close()
-    logger.info(s"Closed $name")
   }
 
   /** Closes the index and removes all its files. */
   def delete() {
     try close() finally FileUtils.deleteRecursive(path.toFile)
-    logger.info(s"Deleted $name")
   }
 
   /** Finds the top hits for a query and sort, starting from an optional position.
@@ -161,14 +152,15 @@ class FSIndex(
     * @return the found documents, sorted according to the supplied [[Sort]] instance
     */
   def search(after: Option[Term], query: Query, sort: Sort, count: Int): DocumentIterator = {
-    logger.debug(
-      s"""Searching in $name
-          | after: $after
-          | query: $query
-          | count: $count
-          | sort :  $sort
-       """.stripMargin)
-    new DocumentIterator(manager, after, mergeSort, sort, query, count, fields)
+    val searcher = manager.acquire()
+    val releaseSearcher = () => manager.release(searcher)
+    new DocumentIterator(searcher, releaseSearcher, after, mergeSort, sort, query, count, fields)
+  }
+
+  def searcher: (IndexSearcher, () => Unit) = {
+    val searcher = manager.acquire()
+    val releaseSearcher = () => manager.release(searcher)
+    (searcher, releaseSearcher)
   }
 
   /** Returns the total number of documents in this index.
@@ -176,7 +168,6 @@ class FSIndex(
     * @return the number of documents
     */
   def getNumDocs: Int = {
-    logger.debug(s"Getting $name num docs")
     doWithSearcher(searcher => searcher.getIndexReader.numDocs)
   }
 
@@ -185,7 +176,6 @@ class FSIndex(
     * @return the number of deleted documents
     */
   def getNumDeletedDocs: Int = {
-    logger.debug(s"Getting $name num deleted docs")
     doWithSearcher(searcher => searcher.getIndexReader.numDeletedDocs)
   }
 
@@ -196,10 +186,8 @@ class FSIndex(
     * @param doWait         `true` if the call should block until the operation completes
     */
   def forceMerge(maxNumSegments: Int, doWait: Boolean) {
-    logger.info(s"Merging $name segments to $maxNumSegments")
     writer.forceMerge(maxNumSegments, doWait)
     writer.commit()
-    logger.info(s"Merged $name segments to $maxNumSegments")
   }
 
   /** Optimizes the index forcing merge of all segments that have deleted documents.
@@ -208,16 +196,13 @@ class FSIndex(
     * @param doWait `true` if the call should block until the operation completes
     */
   def forceMergeDeletes(doWait: Boolean) {
-    logger.info(s"Merging $name segments with deletions")
     writer.forceMergeDeletes(doWait)
     writer.commit()
-    logger.info(s"Merged $name segments with deletions")
   }
 
   /** Refreshes the index readers. */
   def refresh() {
     manager.maybeRefreshBlocking()
-    logger.debug(s"Refreshed $name readers")
   }
 }